requirements.txt
opencv-python
pyyaml
open3d
numpy
scipy
scikit-image
matplotlib
loguru
pycpd
openvino